fix narrowing, clangformat, mutex lock, VideoEncoder::get/setFrameRate to float

+++ include/depthai/pipeline/node/VideoEncoder.hpp
@@ -101,7 +101,7 @@ class VideoEncoder : public Node {
      * Sets expected frame rate
      * @param frameRate Frame rate in frames per second
      */
-    void setFrameRate(int frameRate);
+    void setFrameRate(float frameRate);
 
     /// Get rate control mode
     Properties::RateControlMode getRateControlMode() const;
@@ -123,7 +123,7 @@ class VideoEncoder : public Node {
     /// Get input height
     int getHeight() const;
     /// Get frame rate
-    int getFrameRate() const;
+    float getFrameRate() const;
 };
 
 }  // namespace node
