+++ src/pipeline/node/VideoEncoder.cpp
@@ -89,7 +89,7 @@ void VideoEncoder::setQuality(int quality) {
     properties.quality = quality;
 }
 
-void VideoEncoder::setFrameRate(int frameRate) {
+void VideoEncoder::setFrameRate(float frameRate) {
     properties.frameRate = frameRate;
 }
 
@@ -133,7 +133,7 @@ int VideoEncoder::getHeight() const {
     return std::get<1>(getSize());
 }
 
-int VideoEncoder::getFrameRate() const {
+float VideoEncoder::getFrameRate() const {
     return properties.frameRate;
 }
 
@@ -154,7 +154,7 @@ void VideoEncoder::setDefaultProfilePreset(int width, int height, float fps, Vid
         case VideoEncoderProperties::Profile::H264_MAIN:
         case VideoEncoderProperties::Profile::H265_MAIN: {
             // By default set keyframe frequency to equal fps
-            properties.keyframeFrequency = fps;
+            properties.keyframeFrequency = static_cast<int32_t>(fps);
 
             // Approximate bitrate on input w/h and fps
             constexpr float ESTIMATION_FPS = 30.0f;
@@ -164,16 +164,16 @@ void VideoEncoder::setDefaultProfilePreset(int width, int height, float fps, Vid
             const int pixelArea = width * height;
             if(pixelArea <= 1280 * 720 * AREA_MUL) {
                 // 720p
-                setBitrate((4000 * 1000 / ESTIMATION_FPS) * fps);
+                setBitrate(static_cast<int>((4000 * 1000 / ESTIMATION_FPS) * fps));
             } else if(pixelArea <= 1920 * 1080 * AREA_MUL) {
                 // 1080p
-                setBitrate((8500 * 1000 / ESTIMATION_FPS) * fps);
+                setBitrate(static_cast<int>((8500 * 1000 / ESTIMATION_FPS) * fps));
             } else if(pixelArea <= 2560 * 1440 * AREA_MUL) {
                 // 1440p
-                setBitrate((14000 * 1000 / ESTIMATION_FPS) * fps);
+                setBitrate(static_cast<int>((14000 * 1000 / ESTIMATION_FPS) * fps));
             } else {
                 // 4K
-                setBitrate((20000 * 1000 / ESTIMATION_FPS) * fps);
+                setBitrate(static_cast<int>((20000 * 1000 / ESTIMATION_FPS) * fps));
             }
         } break;
